requirements.txt
streamlit
pyserial
google-generativeai
pandas
plotly
numpy
python-dotenv
pybluez